Refactor to pull some logic into the individual wrappers.
