+++ .github/workflows/scripts/nugetSlackNotifications/packageInfo.json
@@ -156,5 +156,8 @@
     },
     {
         "packageName": "stackexchange.redis"
+    },
+    {
+        "packageName": "opensearch.client"
     }
 ]
