+++ src/Agent/NewRelic/Agent/Extensions/Providers/Wrapper/Elasticsearch/RequestWrapper.cs
@@ -1,6 +1,7 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+using System;
 using System.Threading.Tasks;
 using NewRelic.Agent.Api;
 using NewRelic.Agent.Extensions.Helpers;
@@ -15,6 +16,8 @@ namespace NewRelic.Providers.Wrapper.Elasticsearch
         private const string WrapperName = "ElasticsearchRequestWrapper";
         private const int RequestParamsIndex = 3;
         private const int RequestParamsIndexAsync = 4;
+
+        private static Func<object, object> _apiCallDetailsGetter;
 
         public override DatastoreVendor Vendor => DatastoreVendor.Elasticsearch;
 
@@ -56,7 +59,8 @@ namespace NewRelic.Providers.Wrapper.Elasticsearch
                     }
                     var responseGetter = GetRequestResponseFromGeneric.GetOrAdd(responseTask.GetType(), t => VisibilityBypasser.Instance.GeneratePropertyAccessor<object>(t, "Result"));
                     var response = responseGetter(responseTask);
-                    TryProcessResponse(agent, transaction, response, segment);
+                    _apiCallDetailsGetter ??= GetApiCallDetailsGetterFromResponse(response);
+                    TryProcessResponse(agent, transaction, response, segment, _apiCallDetailsGetter);
                 }
             }
             else
@@ -64,7 +68,8 @@ namespace NewRelic.Providers.Wrapper.Elasticsearch
                 return Delegates.GetDelegateFor<object>(
                         onSuccess: response =>
                         {
-                            TryProcessResponse(agent, transaction, response, segment);
+                            _apiCallDetailsGetter ??= GetApiCallDetailsGetterFromResponse(response);
+                            TryProcessResponse(agent, transaction, response, segment, _apiCallDetailsGetter);
                         },
                         onFailure: exception =>
                         {
@@ -72,6 +77,16 @@ namespace NewRelic.Providers.Wrapper.Elasticsearch
                             segment.End(exception);
                         });
             }
+        }
+
+        private static Func<object, object> GetApiCallDetailsGetterFromResponse(object response)
+        {
+            var typeOfResponse = response.GetType();
+            var responseAssemblyName = typeOfResponse.Assembly.FullName;
+            var apiCallDetailsPropertyName = responseAssemblyName.StartsWith("Elastic.Clients.Elasticsearch")
+                ? "ApiCallDetails" : "ApiCall";
+
+            return VisibilityBypasser.Instance.GeneratePropertyAccessor<object>(responseAssemblyName, typeOfResponse.FullName, apiCallDetailsPropertyName);
         }
     }
 }
